pyproject.toml
[tool.poetry]
name = "django-spam-service-2024"
version = "0.1.0"
description = ""
authors = ["Евгений Ермак <144560375+EvgeniiErmak@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.9"
django-apscheduler = "^0.6.2"
django-cron = "^0.6.0"
psycopg2-binary = "^2.9.9"
ipython = "^8.18.1"
django-debug-toolbar = "^3.2.4"
faker = "^21.0.0"
django = "4.2"
bootstrap4 = "^0.1.0"
django-crispy-forms = "1.9.2"
pillow = "9.2.0"
secure-smtplib = "^0.1.1"
certifi = "2022.9.24"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
cryptography = "38.0.1"
defusedxml = "0.7.1"
idna = "3.4"
oauthlib = "3.2.2"
pycparser = "2.21"
pyjwt = "2.6.0"
python-dotenv = "0.19.0"
python3-openid = "3.2.0"
pytz = "2022.5"
requests = "2.28.1"
requests-oauthlib = "1.3.1"
social-auth-app-django = "5.0.0"
social-auth-core = "4.3.0"
sqlparse = "0.4.3"
urllib3 = "1.26.12"
asgiref = "^3.7.2"
pytils = "^0.4.1"
tasks = "^2.8.0"
django-bootstrap-datepicker-plus = "^5.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
